+++ tests/src/swap_transactions_tests.rs
@@ -78,7 +78,7 @@ async fn test_swap_fully_onchain() {
         )
         .unwrap();
 
-    let expected_output_notes: Vec<Note> = tx_request.expected_output_notes().cloned().collect();
+    let expected_output_notes: Vec<Note> = tx_request.expected_output_own_notes();
     let expected_payback_note_details: Vec<NoteDetails> =
         tx_request.expected_future_notes().cloned().map(|(n, _)| n).collect();
     assert_eq!(expected_output_notes.len(), 1);
@@ -240,7 +240,7 @@ async fn test_swap_private() {
         )
         .unwrap();
 
-    let expected_output_notes: Vec<Note> = tx_request.expected_output_notes().cloned().collect();
+    let expected_output_notes: Vec<Note> = tx_request.expected_output_own_notes();
     let expected_payback_note_details =
         tx_request.expected_future_notes().cloned().map(|(n, _)| n).collect::<Vec<_>>();
     assert_eq!(expected_output_notes.len(), 1);