+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 
 * [BREAKING] Renamed `miden-cli` crate to `miden-client-cli`, and the `miden` executable to `miden-client` (#960).
 * [BREAKING] Merged `concurrent` feature with `std` (#974).
+* [BREAKING] Changed `TransactionRequest` to use expected output recipients instead of output notes (#976).
 
 ### Features